Fix SHA256 on 32-bit builds (undefined 32-bit shift in length padding)
sha256::final computed the message bit-length as (m_tot_len + m_len) << 3 where both members are size_t. On 32-bit (Win32) size_t is 32-bit, so 'len_b >> 32' when packing the high length word is a shift >= type width = undefined behavior; on x86 the shift count masks to 0 and returns len_b unchanged, writing a wrong high length word into the SHA-256 padding and producing an incorrect digest. x64 (size_t = 64-bit) was unaffected. Compute len_b as uint64_t so the padding length is always 64-bit. Fixes the Win32 'Should calc Hashes' SHA256 test failure (337/337 on both Win32 and x64).

+++ src/crypto_sha.cpp
@@ -384,7 +384,7 @@ void sha256::update(const df::cspan cs)
 void sha256::final(uint8_t* digest)
 {
 	const auto block_nb = 1 + (SHA224_256_BLOCK_SIZE - 9 < m_len % SHA224_256_BLOCK_SIZE);
-	const auto len_b = (m_tot_len + m_len) << 3;
+	const auto len_b = static_cast<uint64_t>(m_tot_len + m_len) << 3;
 	const auto pm_len = block_nb << 6;
 
 	memset(m_block + m_len, 0, pm_len - m_len);
